+++ .gitattributes
@@ -1,0 +1,22 @@
+# Set the default behavior, in case people don't have core.autocrlf set.
+* text eol=lf
+
+# Explicitly declare text files you want to always be normalized and converted
+# to native line endings on checkout.
+#*.c text
+#*.h text
+
+# Declare files that will always have CRLF line endings on checkout.
+#*.sln text eol=crlf
+
+# Denote all files that are truly binary and should not be modified.
+*.zip binary
+*.7z binary
+*.png binary
+*.jpg binary
+*.otf binary
+*.eot binary
+*.svg text
+*.ttf binary
+*.woff binary
+*.woff2 binary
