+++ README.md
@@ -1,2 +1,10 @@
 # iota-faucet
 Get 1 free IOTA
+
+git clone this repo
+cd iota-faucet
+
+npm install
+npm start
+
+http://localhost:3000
