+++ public/index.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+<meta charset="utf-8" />
+<title>IOTA FAUCET</title>
+<link rel="stylesheet" href="iota.css" type="text/css" />
+<!--[if IE]>
+  <script src="http://html5shiv.googlecode.com/svn/trunk/html5.js"></script><![endif]-->
+<!--[if lte IE 7]>
+  <script src="js/IE8.js" type="text/javascript"></script><![endif]-->
+<!--[if lt IE 7]>
+  <link rel="stylesheet" type="text/css" media="all" href="css/ie6.css"/><![endif]-->
+</head>
+<body id="index" class="home">
+<div class="container">
+<p style="text-align: center"><img src="iota.png" width="250" style=" margin: 20px auto;"/></p>
+<h1 style="text-align: center; color:white; font-family:'Gill Sans', 'Gill Sans MT', 'Myriad Pro', 'DejaVu Sans Condensed', Helvetica, Arial, 'sans-serif' "> - IOTA Faucet - </h1>
+<h2 style="text-align: center; color:white;">Current supply <span class="label label-primary" id="iota__balance">0</span> </h2>	<h3 style="text-align: center; color:white;">Enter your IOTA address to receive 1 IOTA</h3>
+<div class="input-group">
+<input id="sendIotaInput" class="input" type="text" placeholder="IOTA address" name="address"/>
+<div id="sendIotaBtn" class="button" type="button">Send it</div>
+</div>
+</div>
+</body>
+<script
+  src="https://code.jquery.com/jquery-3.2.1.slim.min.js"
+  integrity="sha256-k2WSCIexGzOj3Euiig+TlR8gA0EmPjuc79OEeY5L45g="
+  crossorigin="anonymous">
+</script>
+<script src="/socket.io/socket.io.js"></script>
+<script src="./js/main.js"></script>
+</html>
